requirements.txt
annotated-types                    0.7.0
anyio                              4.10.0
asttokens                          3.0.0
attrs                              25.3.0
beautifulsoup4                     4.13.4
blinker                            1.9.0
cached-property                    2.0.1
certifi                            2025.8.3
cffi                               1.17.1
chardet                            3.0.4
charset-normalizer                 3.4.3
click                              8.2.1
colorama                           0.4.6
comm                               0.2.3
comtypes                           1.4.11
cryptography                       44.0.3
dataclasses-json                   0.6.7
debugpy                            1.8.16
decorator                          5.2.1
defusedxml                         0.7.1
distro                             1.9.0
dnspython                          2.7.0
docstring_parser_fork              0.0.5
docutils                           0.21.2
et_xmlfile                         2.0.0
exceptiongroup                     1.3.0
exchangelib                        5.5.1
executing                          2.2.0
fire                               0.7.1
Flask                              3.1.2
fonttools                          4.59.1
fpdf2                              2.7.5
furl                               2.1.4
graphviz                           0.13.2
greenlet                           3.2.4
h11                                0.16.0
holidays                           0.45
htmldocx                           0.0.6
httpcore                           1.0.9
httpx                              0.28.1
idna                               3.10
importlib_metadata                 8.7.0
inflection                         0.5.1
ipykernel                          6.30.1
ipython                            9.4.0
ipython_pygments_lexers            1.1.1
isodate                            0.7.2
itsdangerous                       2.2.0
java-access-bridge-wrapper         1.2.0
jedi                               0.19.2
Jinja2                             3.1.6
jsonpath-ng                        1.7.0
jsonschema                         4.25.1
jsonschema-specifications          2025.4.1
jupyter_client                     8.6.3
jupyter_core                       5.8.1
lxml                               6.0.1
MarkupSafe                         3.0.2
marshmallow                        3.26.1
matplotlib-inline                  0.1.7
more-itertools                     10.7.0
mss                                6.1.0
mypy_extensions                    1.1.0
nest_asyncio                       1.6.0
netsuitesdk                        1.24.0
notifiers                          1.3.6
numpy                              2.3.2
O365                               2.0.26
oauthlib                           3.3.1
openpyxl                           3.1.5
orderedmultidict                   1.0.1
outcome                            1.3.0.post0
packaging                          24.2
pandas                             2.3.2
parso                              0.8.5
pdfminer.six                       20221105
pendulum                           3.1.0
pickleshare                        0.7.5
pillow                             10.4.0
pip                                25.2
platformdirs                       4.3.8
playwright                         1.54.0
ply                                3.11
prompt_toolkit                     3.0.51
psutil                             5.9.8
pure_eval                          0.2.3
pyairtable                         3.2.0
pycparser                          2.22
pydantic                           2.11.7
pydantic_core                      2.33.2
pyee                               13.0.0
Pygments                           2.19.2
PyJWT                              2.10.1
pynput-robocorp-fork               5.0.0
pyotp                              2.9.0
pypdf                              3.17.4
pyperclip                          1.9.0
PySocks                            1.7.1
pyspnego                           0.11.2
python-dateutil                    2.9.0.post0
python-docx                        0.8.11
python-dotenv                      1.1.1
pytz                               2025.2
pytz-deprecation-shim              0.1.0.post0
pywin32                            306
pywinauto                          0.6.9
PyYAML                             6.0.2
pyzmq                              27.0.2
RapidFuzz                          3.13.0
referencing                        0.36.2
requests                           2.32.3
requests-file                      2.1.0
requests_ntlm                      1.3.0
requests-oauthlib                  1.3.1
requests-toolbelt                  1.0.0
retell-sdk                         4.45.0
robocorp                           3.0.0
robocorp-browser                   2.3.5
robocorp-log                       3.0.1
robocorp-storage                   1.0.5
robocorp-tasks                     4.0.0
robocorp-vault                     1.3.9
robocorp-workitems                 1.4.7
robotframework                     7.3.2
robotframework-pythonlibcore       4.4.1
robotframework-requests            0.9.7
robotframework-sapguilibrary       1.1
robotframework-seleniumlibrary     6.7.1
robotframework-seleniumtestability 2.1.0
rpaframework                       30.0.2
rpaframework-core                  12.0.2
rpaframework-pdf                   8.0.1
rpaframework-windows               9.0.0
rpds-py                            0.27.0
selenium                           4.15.2
setuptools                         80.9.0
simple-salesforce                  1.12.9
six                                1.17.0
smartsheet-python-sdk              3.0.2
sniffio                            1.3.1
sortedcontainers                   2.4.0
soupsieve                          2.7
sspilib                            0.3.1
stack_data                         0.6.3
stringcase                         1.2.0
tenacity                           8.5.0
termcolor                          3.1.0
tornado                            6.5.2
traitlets                          5.14.3
trio                               0.30.0
trio-websocket                     0.12.2
tweepy                             3.10.0
typing_extensions                  4.14.1
typing-inspect                     0.9.0
typing-inspection                  0.4.1
tzdata                             2025.2
tzlocal                            4.3.1
uiautomation                       2.0.29
urllib3                            2.5.0
validators                         0.34.0
wcwidth                            0.2.13
webdriver-manager                  4.0.2
Werkzeug                           3.1.3
wheel                              0.45.1
wrapt                              1.17.3
wsproto                            1.2.0
xlrd                               2.0.2
xlutils                            2.0.0
xlwt                               1.3.0
zeep                               4.3.1
zipp                               3.23.0
